refactor precompile dependence
