+++ gnovm/cmd/gno/util.go
@@ -1,8 +1,7 @@
-package precompile
+package main
 
 import (
 	"fmt"
-	"go/ast"
 	"io"
 	"io/fs"
 	"os"
@@ -12,104 +11,19 @@ import (
 	"time"
 
 	"github.com/gnolang/gno/gnovm/pkg/gnoenv"
+	"github.com/gnolang/gno/gnovm/pkg/precompile"
 )
 
-func IsGnoFile(f fs.DirEntry) bool {
-	name := f.Name()
-	return !strings.HasPrefix(name, ".") && strings.HasSuffix(name, ".gno") && !f.IsDir()
-}
-
-func IsFileExist(path string) bool {
+func isFileExist(path string) bool {
 	_, err := os.Stat(path)
 	return err == nil
-}
-
-func GnoFilesFromArgs(args []string) ([]string, error) {
-	fmt.Println("---GnoFilesFromArgs, args: ", args)
-	paths := []string{}
-	for _, arg := range args {
-		info, err := os.Stat(arg)
-		if err != nil {
-			return nil, fmt.Errorf("invalid file or package path: %w", err)
-		}
-		if !info.IsDir() {
-			curpath := arg
-			paths = append(paths, curpath)
-		} else {
-			err = filepath.WalkDir(arg, func(curpath string, f fs.DirEntry, err error) error {
-				if err != nil {
-					return fmt.Errorf("%s: walk dir: %w", arg, err)
-				}
-
-				if !IsGnoFile(f) {
-					return nil // skip
-				}
-				paths = append(paths, curpath)
-				return nil
-			})
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return paths, nil
-}
-
-func GnoPackagesFromArgs(args []string) ([]string, error) {
-	paths := []string{}
-	for _, arg := range args {
-		info, err := os.Stat(arg)
-		if err != nil {
-			return nil, fmt.Errorf("invalid file or package path: %w", err)
-		}
-		if !info.IsDir() {
-			paths = append(paths, arg)
-		} else {
-			// if the passed arg is a dir, then we'll recursively walk the dir
-			// and look for directories containing at least one .gno file.
-
-			visited := map[string]bool{} // used to run the builder only once per folder.
-			err = filepath.WalkDir(arg, func(curpath string, f fs.DirEntry, err error) error {
-				if err != nil {
-					return fmt.Errorf("%s: walk dir: %w", arg, err)
-				}
-				if f.IsDir() {
-					return nil // skip
-				}
-				if !IsGnoFile(f) {
-					return nil // skip
-				}
-
-				parentDir := filepath.Dir(curpath)
-				if _, found := visited[parentDir]; found {
-					return nil
-				}
-				visited[parentDir] = true
-
-				pkg := parentDir
-				if !filepath.IsAbs(parentDir) {
-					// cannot use path.Join or filepath.Join, because we need
-					// to ensure that ./ is the prefix to pass to go build.
-					// if not absolute.
-					pkg = "./" + parentDir
-				}
-
-				paths = append(paths, pkg)
-				return nil
-			})
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return paths, nil
 }
 
 // TargetsFromPatterns returns a list of target paths that match the patterns.
 // Each pattern can represent a file or a directory, and if the pattern
 // includes "/...", the "..." is treated as a wildcard, matching any string.
 // Intended to be used by gno commands such as `gno test`.
-func TargetsFromPatterns(patterns []string) ([]string, error) {
+func targetsFromPatterns(patterns []string) ([]string, error) {
 	paths := []string{}
 	for _, p := range patterns {
 		var match func(string) bool
@@ -146,7 +60,7 @@ func TargetsFromPatterns(patterns []string) ([]string, error) {
 				return fmt.Errorf("%s: walk dir: %w", dirToSearch, err)
 			}
 			// Skip directories and non ".gno" files.
-			if f.IsDir() || !IsGnoFile(f) {
+			if f.IsDir() || !precompile.IsGnoFile(f) {
 				return nil
 			}
 
@@ -188,35 +102,22 @@ func matchPattern(pattern string) func(name string) bool {
 	}
 }
 
-func FmtDuration(d time.Duration) string {
+func fmtDuration(d time.Duration) string {
 	return fmt.Sprintf("%.2fs", d.Seconds())
 }
 
 // makeTestGoMod creates the temporary go.mod for test
-func MakeTestGoMod(path string, packageName string, goversion string) error {
+func makeTestGoMod(path string, packageName string, goversion string) error {
 	content := fmt.Sprintf("module %s\n\ngo %s\n", packageName, goversion)
 	return os.WriteFile(path, []byte(content), 0o644)
 }
 
-// getPathsFromImportSpec derive and returns ImportPaths
-// without ImportPrefix from *ast.ImportSpec
-func GetPathsFromImportSpec(importSpec []*ast.ImportSpec) (importPaths []ImportPath) {
-	for _, i := range importSpec {
-		path := i.Path.Value[1 : len(i.Path.Value)-1] // trim leading and trailing `"`
-		if strings.HasPrefix(path, ImportPrefix) {
-			res := strings.TrimPrefix(path, ImportPrefix)
-			importPaths = append(importPaths, ImportPath("."+res))
-		}
-	}
-	return
-}
-
-// ResolvePath joins the output dir with relative pkg path
+// resolvePath joins the output dir with relative pkg path
 // e.g
 // Output Dir: Temp/gno-precompile
 // Pkg Path: ../example/gno.land/p/pkg
 // Returns -> Temp/gno-precompile/example/gno.land/p/pkg
-func ResolvePath(output string, path ImportPath) (string, error) {
+func resolvePath(output string, path precompile.ImportPath) (string, error) {
 	absOutput, err := filepath.Abs(output)
 	if err != nil {
 		return "", err
@@ -228,20 +129,6 @@ func ResolvePath(output string, path ImportPath) (string, error) {
 	pkgPath := strings.TrimPrefix(absPkgPath, gnoenv.RootDir())
 
 	return filepath.Join(absOutput, pkgPath), nil
-}
-
-// WriteDirFile write file to the path and also create
-// directory if needed. with:
-// Dir perm -> 0755; File perm -> 0o644
-func WriteDirFile(pathWithName string, data []byte) error {
-	path := filepath.Dir(pathWithName)
-
-	// Create Dir if not exists
-	if _, err := os.Stat(path); os.IsNotExist(err) {
-		os.MkdirAll(path, 0o755)
-	}
-
-	return os.WriteFile(pathWithName, data, 0o644)
 }
 
 // copyDir copies the dir from src to dst, the paths have to be
@@ -314,7 +201,7 @@ func copyFile(src, dst string) error {
 }
 
 // Adapted from https://yourbasic.org/golang/formatting-byte-size-to-human-readable-format/
-func PrettySize(nb int64) string {
+func prettySize(nb int64) string {
 	const unit = 1000
 	if nb < unit {
 		return fmt.Sprintf("%d", nb)
